moves queries to separate module
