+++ services/catalog/src/simcore_service_catalog/db/repositories/services.py
@@ -1,9 +1,9 @@
 import datetime
+import itertools
 import logging
 from collections import defaultdict
 from collections.abc import Iterable
 from dataclasses import dataclass
-from itertools import chain
 from typing import Any, cast
 
 import packaging.version
@@ -19,69 +19,27 @@ from models_library.services_db import ServiceAccessRightsAtDB, ServiceMetaDataA
 from models_library.users import GroupID, UserID
 from psycopg2.errors import ForeignKeyViolation
 from pydantic import PositiveInt, ValidationError
-from simcore_postgres_database.models.groups import user_to_groups
 from simcore_postgres_database.utils_services import create_select_latest_services_query
 from sqlalchemy import literal_column
-from sqlalchemy.dialects.postgresql import ARRAY, INTEGER, array_agg
 from sqlalchemy.dialects.postgresql import insert as pg_insert
 from sqlalchemy.sql import and_, or_
-from sqlalchemy.sql.expression import func, tuple_
-from sqlalchemy.sql.selectable import Select
+from sqlalchemy.sql.expression import tuple_
 
 from ...models.services_specifications import ServiceSpecificationsAtDB
 from ..tables import services_access_rights, services_meta_data, services_specifications
 from ._base import BaseRepository
+from ._services_sql import (
+    AccessRightsClauses,
+    compose_access_rights_clause,
+    list_services_stmt,
+    list_services_with_history_stmt,
+    total_count_stmt,
+)
 
 _logger = logging.getLogger(__name__)
 
 
-def _make_list_services_statement(
-    *,
-    gids: list[int] | None = None,
-    execute_access: bool | None = None,
-    write_access: bool | None = None,
-    combine_access_with_and: bool | None = True,
-    product_name: str | None = None,
-) -> Select:
-    stmt = sa.select(services_meta_data)
-    if gids or execute_access or write_access:
-        conditions = []
-
-        # access rights
-        logic_operator = and_ if combine_access_with_and else or_
-        default = bool(combine_access_with_and)
-
-        access_query_part = logic_operator(
-            services_access_rights.c.execute_access if execute_access else default,
-            services_access_rights.c.write_access if write_access else default,
-        )
-        conditions.append(access_query_part)
-
-        # on groups
-        if gids:
-            conditions.append(
-                or_(*[services_access_rights.c.gid == gid for gid in gids])
-            )
-
-        # and product name
-        if product_name:
-            conditions.append(services_access_rights.c.product_name == product_name)
-
-        stmt = (
-            sa.select(
-                [services_meta_data],
-            )
-            .distinct(services_meta_data.c.key, services_meta_data.c.version)
-            .select_from(services_meta_data.join(services_access_rights))
-        )
-        if conditions:
-            stmt = stmt.where(and_(*conditions))
-        stmt = stmt.order_by(services_meta_data.c.key, services_meta_data.c.version)
-
-    return stmt
-
-
-def _is_newer(
+def is_newer(
     old: ServiceSpecificationsAtDB | None,
     new: ServiceSpecificationsAtDB,
 ) -> bool:
@@ -91,151 +49,16 @@ def _is_newer(
     )
 
 
-def _merge_specs(
+def merge_specs(
     everyone_spec: ServiceSpecificationsAtDB | None,
     team_specs: dict[GroupID, ServiceSpecificationsAtDB],
     user_spec: ServiceSpecificationsAtDB | None,
 ) -> dict[str, Any]:
     merged_spec = {}
-    for spec in chain([everyone_spec], team_specs.values(), [user_spec]):
+    for spec in itertools.chain([everyone_spec], team_specs.values(), [user_spec]):
         if spec is not None:
             merged_spec.update(spec.dict(include={"sidecar", "service"}))
     return merged_spec
-
-
-def _version(column_or_value):
-    # converts version value string to array[integer] that can be compared
-    # i.e. '1.2.3' -> [1, 2, 3]
-    return sa.func.string_to_array(column_or_value, ".").cast(ARRAY(INTEGER))
-
-
-class AccessRightsClauses:
-    can_execute = services_access_rights.c.execute_access
-    can_read = (
-        services_access_rights.c.execute_access | services_access_rights.c.write_access
-    )
-    can_edit = services_access_rights.c.write_access
-    is_owner = (
-        services_access_rights.c.execute_access & services_access_rights.c.write_access
-    )
-
-
-def _compose_access_rights_clause(
-    product_name: ProductName,
-    user_id: UserID,
-    access_clause: sa.sql.ClauseElement | None = None,
-) -> sa.sql.ClauseElement:
-    conditions: list[sa.sql.ClauseElement] = [
-        services_access_rights.c.product_name == product_name,
-        user_to_groups.c.uid == user_id,
-    ]
-    if access_clause is not None:
-        conditions.append(access_clause)
-    return and_(*conditions)
-
-
-_join_service_metadata_and_access_rights = services_meta_data.join(
-    services_access_rights,
-    (services_meta_data.c.key == services_access_rights.c.key)
-    & (services_meta_data.c.version == services_access_rights.c.version),
-).join(
-    user_to_groups,
-    services_access_rights.c.gid == user_to_groups.c.gid,
-)
-
-
-def _list_services_key_version_stmt(access_rights: sa.sql.ClauseElement):
-    return (
-        sa.select(services_meta_data.c.key, services_meta_data.c.version)
-        .select_from(_join_service_metadata_and_access_rights)
-        .where(access_rights)
-        .distinct()  # Multiple gid of the same uid can have access to the same (key, version). Therefore they apper repeated
-        .order_by(
-            services_meta_data.c.key,
-            sa.desc(_version(services_meta_data.c.version)),  # latest version first
-        )
-    )
-
-
-def _total_count_stmt(access_rights: sa.sql.ClauseElement):
-    return (
-        sa.select(func.count(sa.distinct(services_meta_data.c.keys)))
-        .select_from(_join_service_metadata_and_access_rights)
-        .where(access_rights)
-    )
-
-
-def _list_services_with_history_stmt(
-    *, access_rights: sa.sql.ClauseElement, limit: int | None, offset: int | None
-):
-    #
-    # Common Table Expression (CTE) to select distinct service names with pagination.
-    # This allows limiting the subquery to the required pagination instead of paginating at the last query.
-    # SEE https://learnsql.com/blog/cte-with-examples/
-    #
-    cte = (
-        sa.select(services_meta_data.c.key)
-        .distinct()
-        .select_from(
-            services_meta_data.join(
-                services_access_rights,
-                (services_meta_data.c.key == services_access_rights.c.key)
-                & (services_meta_data.c.version == services_access_rights.c.version),
-            ).join(
-                user_to_groups,
-                services_access_rights.c.gid == user_to_groups.c.gid,
-            )
-        )
-        .where(access_rights)
-        .order_by(services_meta_data.c.key)  # NOTE: add here the order
-        .limit(limit)
-        .offset(offset)
-        .cte("paginated_services")
-    )
-
-    subquery = (
-        sa.select(
-            services_meta_data.c.key,
-            services_meta_data.c.version,
-            services_meta_data.c.deprecated,
-            services_meta_data.c.created,
-        )
-        .select_from(
-            services_meta_data.join(
-                cte,
-                services_meta_data.c.key == cte.c.key,
-            )
-            # joins because access-rights might change per version
-            .join(
-                services_access_rights,
-                (services_meta_data.c.key == services_access_rights.c.key)
-                & (services_meta_data.c.version == services_access_rights.c.version),
-            ).join(
-                user_to_groups,
-                services_access_rights.c.gid == user_to_groups.c.gid,
-            )
-        )
-        .where(access_rights)
-        .order_by(
-            services_meta_data.c.key,
-            sa.desc(_version(services_meta_data.c.version)),  # latest version first
-        )
-        .subquery()
-    )
-
-    return sa.select(
-        subquery.c.key,
-        array_agg(
-            func.json_build_object(
-                "version",
-                subquery.c.version,
-                "deprecated",
-                subquery.c.deprecated,
-                "created",
-                subquery.c.created,
-            )
-        ).label("history"),
-    ).group_by(subquery.c.key)
 
 
 @dataclass
@@ -270,7 +93,7 @@ class ServicesRepository(BaseRepository):
             return [
                 ServiceMetaDataAtDB.from_orm(row)
                 async for row in await conn.stream(
-                    _make_list_services_statement(
+                    list_services_stmt(
                         gids=gids,
                         execute_access=execute_access,
                         write_access=write_access,
@@ -403,14 +226,14 @@ class ServicesRepository(BaseRepository):
         offset: int | None = None,
     ) -> tuple[PositiveInt, list[ServiceHistoryItem]]:
 
-        access_rights = _compose_access_rights_clause(
+        access_rights = compose_access_rights_clause(
             product_name=product_name,
             user_id=user_id,
             access_clause=AccessRightsClauses.can_read,
         )
 
-        stmt_total = _total_count_stmt(access_rights=access_rights)
-        stmt_page = _list_services_with_history_stmt(
+        stmt_total = total_count_stmt(access_rights=access_rights)
+        stmt_page = list_services_with_history_stmt(
             access_rights=access_rights,
             limit=limit,
             offset=offset,
@@ -632,16 +455,16 @@ class ServicesRepository(BaseRepository):
                         continue
                     # filter by group type
                     group = gid_to_group_map[row.gid]
-                    if (group.group_type == GroupTypeInModel.STANDARD) and _is_newer(
+                    if (group.group_type == GroupTypeInModel.STANDARD) and is_newer(
                         teams_specs.get(db_service_spec.gid),
                         db_service_spec,
                     ):
                         teams_specs[db_service_spec.gid] = db_service_spec
-                    elif (group.group_type == GroupTypeInModel.EVERYONE) and _is_newer(
+                    elif (group.group_type == GroupTypeInModel.EVERYONE) and is_newer(
                         everyone_specs, db_service_spec
                     ):
                         everyone_specs = db_service_spec
-                    elif (group.group_type == GroupTypeInModel.PRIMARY) and _is_newer(
+                    elif (group.group_type == GroupTypeInModel.PRIMARY) and is_newer(
                         primary_specs, db_service_spec
                     ):
                         primary_specs = db_service_spec
@@ -653,7 +476,7 @@ class ServicesRepository(BaseRepository):
                         f"{exc}",
                     )
 
-        if merged_specifications := _merge_specs(
+        if merged_specifications := merge_specs(
             everyone_specs, teams_specs, primary_specs
         ):
             return ServiceSpecifications.parse_obj(merged_specifications)
